# authentification_api/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "paths": {
        "/api/register": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Register a new user",
                "operationId": "b93ae58c0225bccb3c5c1bd20026bb84",
                "parameters": [
                    {
                        "name": "name",
                        "in": "query",
                        "description": "User's name",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "email",
                        "in": "query",
                        "description": "User's email",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "password",
                        "in": "query",
                        "description": "User's password",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "User registered successfully"
                    },
                    "422": {
                        "description": "Validation errors"
                    }
                }
            }
        },
        "/api/login": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Authenticate user and generate passport token",
                "operationId": "afd167661891a2c4c1b49a04643b9223",
                "parameters": [
                    {
                        "name": "email",
                        "in": "query",
                        "description": "User's email",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "password",
                        "in": "query",
                        "description": "User's password",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Login successful"
                    },
                    "401": {
                        "description": "Invalid credentials"
                    }
                }
            }
        },
        "/logout": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Logout user",
                "operationId": "716ee9fdd4df72c5cb7cd983a3a5c2ef",
                "responses": {
                    "200": {
                        "description": "Successful logout"
                    },
                    "400": {
                        "description": "Invalid token"
                    }
                }
            }
        },
        "/api/permissions": {
            "get": {
                "tags": [
                    "Permissions"
                ],
                "summary": "List all permissions",
                "operationId": "2d96d24ccf9aedd18a6a1855536115b7",
                "parameters": [
                    {
                        "name": "read_permissions",
                        "in": "query",
                        "description": "Indicates the action of reading permissions",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "criteria",
                        "in": "query",
                        "description": "Some optional other parameter",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Returns a list of permissions",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Permission"
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Error: Bad request. When required parameters were not supplied."
                    }
                }
            },
            "post": {
                "tags": [
                    "Permissions"
                ],
                "summary": "Permissions data",
                "operationId": "b8001f26f827c8ac85ac9fe8aaf9669c",
                "parameters": [
                    {
                        "name": "name",
                        "in": "query",
                        "description": "",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "criteria",
                        "in": "query",
                        "description": "Some optional other parameter",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Returns some sample category things",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "400": {
                        "description": "Error: Bad request. When required parameters were not supplied."
                    }
                }
            }
        },
        "/api/permissions/{id}": {
            "delete": {
                "tags": [
                    "Permissions"
                ],
                "summary": "Delete a permission",
                "operationId": "6d0e311749fe203258605ab5c093da27",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the permission to delete",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Permission deleted successfully"
                    },
                    "404": {
                        "description": "Permission not found"
                    }
                }
            }
        },
        "/api/roles": {
            "get": {
                "tags": [
                    "Roles"
                ],
                "summary": "Read roles",
                "operationId": "f9ef6a6dc70ceaff2dbd755f9f66fe15",
                "parameters": [
                    {
                        "name": "read roles",
                        "in": "query",
                        "description": "en tant qu'admin, je peux afficher les roles",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "criteria",
                        "in": "query",
                        "description": "Some optional other parameter",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Returns some sample category things",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "400": {
                        "description": "Error: Bad request. When required parameters were not supplied."
                    }
                }
            },
            "post": {
                "tags": [
                    "Roles"
                ],
                "summary": "Roles data",
                "operationId": "c01c4c899efe8db3f91aacc466c55a4a",
                "parameters": [
                    {
                        "name": "name",
                        "in": "query",
                        "description": "",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "criteria",
                        "in": "query",
                        "description": "Some optional other parameter",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Returns some sample category things",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "400": {
                        "description": "Error: Bad request. When required parameters were not supplied."
                    }
                }
            }
        },
        "/api/roles/{id}": {
            "delete": {
                "tags": [
                    "Roles"
                ],
                "summary": "Delete roles",
                "operationId": "6c08c94c32c9933602d7026f71669777",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the permission to delete",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Permission deleted successfully"
                    },
                    "404": {
                        "description": "Permission not found"
                    }
                }
            }
        },
...
    },
...
}